Add forecast models and datas to ui
