+++ pubspec.yaml
@@ -18,10 +18,10 @@ environment:
 
 dependencies:
 
-  dio: 2.2.2
+  #dio: 2.2.2
   json_serializable: ^3.2.2
   json_annotation: 3.0.0
-  retrofit: any
+  retrofit: 1.0.1
   logger: 0.7.0  #for logging purpose
   provider: 3.1.0
